Separate gameObjects with pure models
