+++ client-src/models/collisionCategory.ts
