+++ client-src/scenes/MainScene.ts
@@ -5,13 +5,13 @@ import { GameObjects } from 'phaser';
 import { preload as _preload, setUpAudio } from '../assets';
 import { config, ItemType } from '../config/config';
 import { BASE_LINE_WIDTH, BULLET_SPEED, DEBUG_DISABLE_SPAWNING, DEBUG_PHYSICS, PHYSICS_FRAME_SIZE, PHYSICS_MAX_FRAME_CATCHUP, PIXEL_TO_METER, PLAYER_MOVE_SPEED, SPAWN_DELAY, SPAWN_INTERVAL, TANK_CHASE_ITEM_RANGE, TANK_SPEED, WORLD_HEIGHT, WORLD_WIDTH } from '../constants';
-import { Bullet } from '../entities/Bullet';
-import { Item } from '../entities/Item';
+import { Bullet } from '../gameObjects/Bullet';
+import { Item } from '../gameObjects/Item';
 // import { Immutable } from '../utils/ImmutableType';
-import { Player } from '../entities/Player';
-import { Tank } from '../entities/Tank';
-import { Team } from '../entities/Team';
-import { UpgradeObject } from '../entities/Upgrade';
+import { Player } from '../gameObjects/Player';
+import { Tank } from '../gameObjects/Tank';
+import { Team } from '../models/Team';
+import { UpgradeObject } from '../models/Upgrade';
 import { IBodyUserData, IFixtureUserData, PhysicsSystem } from '../PhysicsSystem';
 import { HpBar } from '../ui/HpBar';
 import { DistanceMatrix } from '../utils/DistanceMatrix';
